+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * Update the `org.realityforge.guiceyloops` artifact to version `0.110`.
 * Update the `org.realityforge.grim` artifacts to version `0.06`.
 * Update the `org.realityforge.akasha` artifacts to version `0.30`.
+* Add a `AbstractSessionContextImpl` base class that SessionContext implementations can extend.
 
 ### [v6.114](https://github.com/replicant4j/replicant/tree/v6.114) (2021-10-23) · [Full Changelog](https://github.com/spritz/spritz/compare/v6.113...v6.114)